+++ R/gaussian_mechanism.R
@@ -1,4 +1,4 @@
-#' @include mechanisms.R privacy_params.R utils.R
+#' @include mechanisms.R numeric_mechanism.R privacy_params.R utils.R
 NULL
 
 #' An S4 class for the Gaussian mechanism of differential privacy.
@@ -10,10 +10,10 @@ NULL
 #'   Defaults to \code{Inf} for use with \code{sensitivitySampler()}.
 #' @slot target the target non-private function to be privatized, takes lists.
 #'   Defaults to a constant function. Gaussian mechanism assumes functions that
-#'   release numeric vectors of fixed dimension \code{dim}.
+#'   release numeric vectors of fixed dimension \code{dims}.
 #' @slot gammaSensitivity \code{NA_real_} if deactive, or scalar in [0,1)
 #'   indicating that responses must be RDP with specific confidence.
-#' @slot dim positive scalar numeric dimension of responses. Defaults to
+#' @slot dims positive scalar numeric dimension of responses. Defaults to
 #'   \code{NA_integer_} for use with \code{sensitivitySampler()} which can
 #'   probe \code{target} to determine dimension.
 #'
@@ -24,32 +24,14 @@ NULL
 #' @export DPMechGaussian
 #' @exportClass DPMechGaussian
 DPMechGaussian <- setClass("DPMechGaussian",
-  contains = "DPMech",
-  slots = list(dim = "numeric"),
-  prototype = prototype(dim = NA_integer_)
+  contains = "DPMechNumeric"
 )
-
-## A \code{DPMechGaussian} should be constructed with an appropriate dimension.
-setValidity("DPMechGaussian", function(object) {
-  if (!is.na(object@dim) && !.check_integer(object@dim)) {
-    return("DPMechGaussian@dim should be a scalar natural number.")
-  }
-  return(TRUE)
-})
 
 #' @describeIn DPMechGaussian automatically prints the object.
 #' @param object an instance of class \code{DPMech}.
 setMethod("show", "DPMechGaussian", function(object) {
   cat("Gaussian mechanism\n")
-  cat("Sensitivity:", object@sensitivity, "\n")
-  if (is.na(object@gammaSensitivity)) {
-    cat("Sampled sensitivity gamma: NA\n")
-  } else {
-    cat("Sampled sensitivity gamma:", object@gammaSensitivity, "\n")
-  }
-  cat("Response dimension:", object@dim, "\n")
-  cat("Target function: \n")
-  show(object@target)
+  callNextMethod()
 })
 
 #' @describeIn DPMechGaussian releases Gaussian mechanism responses.
@@ -59,11 +41,11 @@ setMethod("show", "DPMechGaussian", function(object) {
 #'   matrix, data frame, numeric/character vector.
 #' @return list with slots per argument, actual privacy parameter; Gaussian
 #'   mechanism response with length of target release:
-#'   \code{privacyParams, sensitivity, dim, target, response}.
+#'   \code{privacyParams, sensitivity, dims, target, response}.
 #' @examples
 #' f <- function(xs) mean(xs)
 #' n <- 100
-#' m <- DPMechGaussian(sensitivity = 1/n, target = f, dim = 1)
+#' m <- DPMechGaussian(sensitivity = 1/n, target = f, dims = 1)
 #' X <- runif(n)
 #' p <- DPParamsDel(epsilon = 1, delta = 0.1)
 #' releaseResponse(m, p, X)
@@ -77,10 +59,10 @@ setMethod("releaseResponse",
     if (!is.numeric(rawR)) {
       stop("Non-private target output non-numeric.")
     }
-    if (is.na(mechanism@dim)) {
-      warning("No expected non-private dim slot set.")
+    if (is.na(mechanism@dims)) {
+      warning("No expected non-private dims slot set.")
     }
-    if (length(rawR) != mechanism@dim) {
+    if (length(rawR) != mechanism@dims) {
       warning("Non-private target output has unexpected dimension.")
     }
     C <- sqrt(2 * log(1.25 / privacyParams@delta))
@@ -95,48 +77,26 @@ setMethod("releaseResponse",
     return(list(
       privacyParams = p,
       sensitivity = mechanism@sensitivity,
-      dim = mechanism@dim,
+      dims = mechanism@dims,
       target = mechanism@target,
       response = R
     ))
   }
 )
 
-#' @describeIn DPMechGaussian measures sensitivity of non-private \code{target}.
-#' @param X1 a privacy-sensitive dataset, list if sensitivity sampler compatible.
-#' @param X2 a privacy-sensitive dataset, list if sensitivity sampler compatible.
-#' @return scalar numeric norm of non-private \code{target} on datasets.
-#' @examples
-#' f <- function(xs) mean(xs)
-#' n <- 100
-#' m <- DPMechGaussian(sensitivity = 1/n, target = f, dim = 1)
-#' X1 <- runif(n)
-#' X2 <- runif(n)
-#' sensitivityNorm(m, X1, X2)
-#' @export
-setMethod("sensitivityNorm",
-  signature(mechanism = "DPMechGaussian",
-            X1 = "ANY",
-            X2 = "ANY"),
-  function(mechanism, X1, X2) {
-    rawR1 <- mechanism@target(X1)
-    rawR2 <- mechanism@target(X2)
-    if (!is.numeric(rawR1) || !is.numeric(rawR2)) {
-      stop("Non-private target output non-numeric.")
-    }
-    if (is.na(mechanism@dim)) {
-      warning("No expected dimension set.")
-    } else {
-      if (length(rawR1) != mechanism@dim || length(rawR2) != mechanism@dim) {
-        warning("Non-private target output has unexpected dimension.")
-      }
-    }
-    if (length(rawR1) != length(rawR2)) {
-      stop("Non-private target output dimensions inconsistent.")
-    }
-    if (length(rawR1) == 0) {
-      return(0)
-    }
+#' \code{DPMechGaussian} response space L2 norm.
+#'
+#' Represents the L2 norm on \code{target} responses. For internal use.
+#'
+#' @param object an instance of class \code{\link{DPMechGaussian-class}}.
+#' @param rawR1 a non-private response from \code{target}.
+#' @param rawR2 a non-private response from \code{target}.
+#' @return a non-negative scalar L2 norm between \code{rawR1}, \code{rawR2}.
+setMethod(".numericNorm",
+  signature(object = "DPMechGaussian",
+            rawR1 = "numeric",
+            rawR2 = "numeric"),
+  function(object, rawR1, rawR2) {
     return(.l2norm(rawR1 - rawR2))
   }
 )