+++ tests/testthat/test_ssampler.R
@@ -2,7 +2,7 @@ library(diffpriv)
 context("DPMech sensitivity sampling")
 
 test_that("sensitivitySampler methods check types", {
-  m <- DPMechLaplace(target = function(xs) mean(unlist(xs)), dim = 1)
+  m <- DPMechLaplace(target = function(xs) mean(unlist(xs)), dims = 1)
   P <- function(n) if (n > 1) as.list(rnorm(n)) else rnorm(1)
   expect_error(sensitivitySamplerManual(m, oracle=P, n=1.1, m=100, k=99),
     "integer", ignore.case=TRUE)
@@ -29,7 +29,7 @@ test_that("sensitivitySampler methods check types", {
 })
 
 test_that("sensitivitySampler() returns expected responses", {
-  m <- DPMechLaplace(target = function(xs) mean(unlist(xs)), dim = 1)
+  m <- DPMechLaplace(target = function(xs) mean(unlist(xs)), dims = 1)
   expect_identical(m@sensitivity, Inf)
   expect_true(is.na(m@gammaSensitivity))
   P <- function(n) if (n > 1) as.list(rnorm(n)) else rnorm(1)
@@ -41,7 +41,7 @@ test_that("sensitivitySampler() returns expected responses", {
 
 test_that("Responses post sensitivitySampler() are RDP", {
   dim <- 1
-  m <- DPMechLaplace(target = function(xs) mean(unlist(xs)), dim = dim)
+  m <- DPMechLaplace(target = function(xs) mean(unlist(xs)), dims = dim)
   P <- function(n) if (n > 1) as.list(rnorm(n)) else rnorm(1)
   epsilon <- 0.1
   gamma <- 0.33
@@ -60,21 +60,21 @@ test_that("Responses post sensitivitySampler() are RDP", {
   expect_equal(getGamma(R$privacyParams), gamma)
 
   ## Case: Given m only
-  m <- DPMechLaplace(target = function(xs) mean(unlist(xs)), dim = dim)
+  m <- DPMechLaplace(target = function(xs) mean(unlist(xs)), dims = dim)
   m <- sensitivitySampler(m, oracle = P, n = n, m = 50)
   R <- releaseResponse(m, DPParamsEps(epsilon = epsilon), rnorm(n))
   expect_is(R$privacyParams, "DPParamsGam")
 
   ## Case: Given both m, gamma
-  m <- DPMechLaplace(target = function(xs) mean(unlist(xs)), dim = dim)
+  m <- DPMechLaplace(target = function(xs) mean(unlist(xs)), dims = dim)
   m <- sensitivitySampler(m, oracle = P, n = n, gamma = gamma, m = 50)
   R <- releaseResponse(m, DPParamsEps(epsilon = epsilon), rnorm(n))
   expect_is(R$privacyParams, "DPParamsGam")
 })
 
-test_that("sensitivitySampler() determines DPMechLaplace@dim", {
+test_that("sensitivitySampler() determines DPMechLaplace@dims", {
   m <- DPMechLaplace(target = function(xs) apply(xs, 2, mean))
   P <- function(n) matrix(rnorm(3*n), ncol=3)
   m <- sensitivitySampler(m, oracle = P, n = 10, gamma = 0.33)
-  expect_equal(m@dim, 3)
+  expect_equal(m@dims, 3)
 })